requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
flask-cors>=4.0.0
pywebview>=4.4.1
pyinstaller>=6.0.0
simple-websocket>=1.0.0
